Validate against FHIR in BaseClient after extraction is complete
