+++ package.json
@@ -19,6 +19,7 @@
   },
   "license": "Apache-2.0",
   "dependencies": {
+    "ajv": "^6.12.6",
     "axios": "^0.19.0",
     "commander": "^6.2.0",
     "csv-parse": "^4.8.8",
@@ -32,7 +33,6 @@
   },
   "devDependencies": {
     "@types/jest": "^26.0.14",
-    "ajv": "^6.12.6",
     "cross-env": "^7.0.0",
     "eslint": "^6.6.0",
     "eslint-config-airbnb-base": "^14.0.0",
